requirements.txt
yt-dlp>=2024.3.10
google-generativeai>=0.5.0
